Fix a iterator problem

+++ app/src/main/java/com/CHH2000day/navalcreed/modhelper/CustomShipNameHelper.java
@@ -111,7 +111,7 @@ public class CustomShipNameHelper
 		bs.writeUtf8(HOF);
 		while (li.hasNext())
 		{
-			int i=li.next();
+			int i=(Integer)li.next();
 			name = shipnames.getOrDefault(i, name);
 			if (name.equals(empty))
 			{continue;}
